requirements.txt
django>=2.2
mysqlclient>=1.4.2.post1
pillow
django-pure-pagination